updated docs, hide product groups endpoints, change authorisation of users endpoint

+++ src/main/java/shop/holy/v3/ecommerce/service/biz/CategoryService.java
@@ -81,6 +81,9 @@ public class CategoryService {
 
     @Transactional
     public int deleteCategories(long[] ids, boolean isHard) {
+        if (ids.length == 0)
+            return 0;
+
         categoryRepository.deleteProductCategoryByCategoryIdIn(ids);
         if (isHard)
             return categoryRepository.deleteAllByIdIn(ids);
